1
1. Переписан класс GameDrawOnFrames, добавлен класс
gameDrawOnFramesWithRangeValues, соответственно упрощены почти все
классы-наследники GameDrawOnFrames. Переписан класс UnitBitmap, теперь в
нём хранятся координаты на которых нужно рисовать изображение юнита, и
также здоровье юнита. Это позволило исправить очень давнюю ошибку: ходим
королём/солдатом на домик, быстро нажимаем захватить и юнит дублируется:
один появляется на домике, а прежний продолжает двигаться до домика.
2. При старте игры создаются изображения чисел от -100 до 100, это
позволило упростить механизм рисования здоровья юнита и также информации
в gameDrawInfo

+++ src/ru/ancientempires/view/draws/GameDrawRange.java
@@ -106,8 +106,7 @@ public class GameDrawRange extends GameDraw
 		this.rangeEndX = (this.startJ + radius + 1) * GameDraw.A;
 		
 		this.rangePath = new Path();
-		this.rangePath.addRect(this.rangeStartX, this.rangeStartY, this.rangeEndX, this.rangeEndY,
-				Direction.CW);
+		this.rangePath.addRect(this.rangeStartX, this.rangeStartY, this.rangeEndX, this.rangeEndY, Direction.CW);
 	}
 	
 	private void updateFrames(int framesForCell)
@@ -176,8 +175,8 @@ public class GameDrawRange extends GameDraw
 		if (this.radius == this.radiusEnd && this.radiusEnd == this.radiusMin)
 			this.isDrawing = false;
 		
-		int canvasTranslateY = this.gameDraw.offsetY; // canvas.getTranslateY();
-		int canvasTranslateX = this.gameDraw.offsetX; // canvas.getTranslateX();
+		float canvasTranslateY = this.gameDraw.offsetY; // canvas.getTranslateY();
+		float canvasTranslateX = this.gameDraw.offsetX; // canvas.getTranslateX();
 		
 		this.cacheBitmap.eraseColor(Color.TRANSPARENT);
